+++ test/fixtures/parser/bigquery/select_except_replace.sql
@@ -1,0 +1,5 @@
+-- We can call functions when replacing a field
+select
+  * except(foo)
+    replace (concat(fruit, 'berry') as fruit)
+from some_table